+++ src/kusto/azext_kusto/vendored_sdks/kusto/aio/_kusto_management_client.py
@@ -6,16 +6,14 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from typing import TYPE_CHECKING
+from typing import Any, Optional, TYPE_CHECKING
 
-from azure.mgmt.core import ARMPipelineClient
+from azure.mgmt.core import AsyncARMPipelineClient
 from msrest import Deserializer, Serializer
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
-    from typing import Any, Optional
-
-    from azure.core.credentials import TokenCredential
+    from azure.core.credentials_async import AsyncTokenCredential
 
 from ._configuration import KustoManagementClientConfiguration
 from .operations import ClustersOperations
@@ -27,32 +25,32 @@ from .operations import AttachedDatabaseConfigurationsOperations
 from .operations import DataConnectionsOperations
 from .operations import Operations
 from .operations import OperationsResultsOperations
-from . import models
+from .. import models
 
 
 class KustoManagementClient(object):
     """The Azure Kusto management API provides a RESTful set of web services that interact with Azure Kusto services to manage your clusters and databases. The API enables you to create, update, and delete clusters and databases.
 
     :ivar clusters: ClustersOperations operations
-    :vartype clusters: kusto_management_client.operations.ClustersOperations
+    :vartype clusters: kusto_management_client.aio.operations.ClustersOperations
     :ivar cluster_principal_assignments: ClusterPrincipalAssignmentsOperations operations
-    :vartype cluster_principal_assignments: kusto_management_client.operations.ClusterPrincipalAssignmentsOperations
+    :vartype cluster_principal_assignments: kusto_management_client.aio.operations.ClusterPrincipalAssignmentsOperations
     :ivar databases: DatabasesOperations operations
-    :vartype databases: kusto_management_client.operations.DatabasesOperations
+    :vartype databases: kusto_management_client.aio.operations.DatabasesOperations
     :ivar database_principal_assignments: DatabasePrincipalAssignmentsOperations operations
-    :vartype database_principal_assignments: kusto_management_client.operations.DatabasePrincipalAssignmentsOperations
+    :vartype database_principal_assignments: kusto_management_client.aio.operations.DatabasePrincipalAssignmentsOperations
     :ivar scripts: ScriptsOperations operations
-    :vartype scripts: kusto_management_client.operations.ScriptsOperations
+    :vartype scripts: kusto_management_client.aio.operations.ScriptsOperations
     :ivar attached_database_configurations: AttachedDatabaseConfigurationsOperations operations
-    :vartype attached_database_configurations: kusto_management_client.operations.AttachedDatabaseConfigurationsOperations
+    :vartype attached_database_configurations: kusto_management_client.aio.operations.AttachedDatabaseConfigurationsOperations
     :ivar data_connections: DataConnectionsOperations operations
-    :vartype data_connections: kusto_management_client.operations.DataConnectionsOperations
+    :vartype data_connections: kusto_management_client.aio.operations.DataConnectionsOperations
     :ivar operations: Operations operations
-    :vartype operations: kusto_management_client.operations.Operations
+    :vartype operations: kusto_management_client.aio.operations.Operations
     :ivar operations_results: OperationsResultsOperations operations
-    :vartype operations_results: kusto_management_client.operations.OperationsResultsOperations
+    :vartype operations_results: kusto_management_client.aio.operations.OperationsResultsOperations
     :param credential: Credential needed for the client to connect to Azure.
-    :type credential: ~azure.core.credentials.TokenCredential
+    :type credential: ~azure.core.credentials_async.AsyncTokenCredential
     :param subscription_id: Gets subscription credentials which uniquely identify Microsoft Azure subscription. The subscription ID forms part of the URI for every service call.
     :type subscription_id: str
     :param str base_url: Service URL
@@ -61,16 +59,15 @@ class KustoManagementClient(object):
 
     def __init__(
         self,
-        credential,  # type: "TokenCredential"
-        subscription_id,  # type: str
-        base_url=None,  # type: Optional[str]
-        **kwargs  # type: Any
-    ):
-        # type: (...) -> None
+        credential: "AsyncTokenCredential",
+        subscription_id: str,
+        base_url: Optional[str] = None,
+        **kwargs: Any
+    ) -> None:
         if not base_url:
             base_url = 'https://management.azure.com'
         self._config = KustoManagementClientConfiguration(credential, subscription_id, **kwargs)
-        self._client = ARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
+        self._client = AsyncARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
@@ -95,15 +92,12 @@ class KustoManagementClient(object):
         self.operations_results = OperationsResultsOperations(
             self._client, self._config, self._serialize, self._deserialize)
 
-    def close(self):
-        # type: () -> None
-        self._client.close()
+    async def close(self) -> None:
+        await self._client.close()
 
-    def __enter__(self):
-        # type: () -> KustoManagementClient
-        self._client.__enter__()
+    async def __aenter__(self) -> "KustoManagementClient":
+        await self._client.__aenter__()
         return self
 
-    def __exit__(self, *exc_details):
-        # type: (Any) -> None
-        self._client.__exit__(*exc_details)
+    async def __aexit__(self, *exc_details) -> None:
+        await self._client.__aexit__(*exc_details)