+++ src/aks-preview/HISTORY.md
@@ -5,6 +5,7 @@ Release History
 0.5.12
 +++++
 * Add --enable-azure-rbac and --disable-azure-rbac in aks update
+* Support disabling local accounts
 
 0.5.11
 +++++